MobileApp/requirements.txt

Flask
requests
pymongo
python-dotenv
psycopg2
datetime